requirements.txt
agentql==1.8.0
python-dotenv==1.0.0
requests==2.31.0
playwright==1.39.0